[#39] Unify all DashboardMenuControllers in one single file
