+++ csc-cmo-ssapp/code/pages/review-order.html
@@ -93,7 +93,7 @@
 					<!-- Attach Documents-->
 					<div class='row'>
 						<div class='col-12 upload-attachment'>
-							<psk-files-chooser event-name='add-file' view-model='form.attachment'></psk-files-chooser>
+							<psk-files-chooser event-name='add-file' data-view-model='@form.attachment'></psk-files-chooser>
 							<p>Maximum upload file size: 64MB</p>
 						</div>
 						<div class='col-12'>
