+++ csc-services/controllers/impl/index.js
@@ -3,3 +3,4 @@ require("./HistoryModalControllerImpl");
 require("./TableTemplateControllerImpl");
 require("./HeaderControllerImpl");
 require("./NotificationsControllerImpl");
+require("./DashboardMenuControllerImpl");
