+++ csc-sponsor-ssapp/code/pages/newOrder.html
@@ -170,7 +170,7 @@
 					<!-- Attach Documents-->
 					<div class='row'>
 						<div class='col-12 upload-attachment'>
-							<psk-files-chooser data-view-model='form.attachment'
+							<psk-files-chooser data-view-model='@form.attachment'
 							                   event-name='add-file'></psk-files-chooser>
 							<p>Maximum upload file size: 64MB</p>
 						</div>
